Add Jupyter notebook for amount organization
This notebook introduces the problem, and provides some background and
rationale for some of the approaches taken so far, as well as a number of plots
including some interactivity to help explore the parameter space.

Preliminary results are two randomized amount decomposition strategies that
achieve fairly good results from a privacy standpoint. Overheads from the
restriction to standard denominations seem to be significant but not cost
prohibitive, and relatively predictable.

The decomposition solutions per used are optimized with a pluggable cost
functions which currently account for the cost of a solution, but do not
optimize for privacy yet (that is the next step).

In specific transactions input registrations determine a further subset of the
standard denominations to be used for the transaction, improving the likelyhood
of equal amount outputs especially for smaller transactions over using the full
set of standard denominations. The algorithms are written in a way that
knowledge about output amounts could be incorporated but is not required, and
one of the goals is to eventually investigate whether or not this tradeoff is
worth it by simulation.